[Decomp][Unit] Cohesion split: extract aura system to UnitAura.cpp
